+++ Api/demo/src/main/java/com/example/shop/demo/Abuse.java
@@ -1,0 +1,72 @@
+package com.example.shop.demo;
+
+import jakarta.persistence.*;
+
+
+@Entity
+@Table(name = "\"abuses\"")
+public class Abuse {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "message_id")
+    private Message message;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User sender;
+
+    private String reason;
+    private String time;
+    private String description;
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getReason() {
+        return reason;
+    }
+
+    public void setReason(String reason) {
+        this.reason = reason;
+    }
+
+    public String getTime() {
+        return time;
+    }
+
+    public void setTime(String time) {
+        this.time = time;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Message getMessage() {
+        return message;
+    }
+
+    public void setMessage(Message message) {
+        this.message = message;
+    }
+
+    public User getSender() {
+        return sender;
+    }
+
+    public void setSender(User sender) {
+        this.sender = sender;
+    }
+}