+++ Api/demo/src/main/java/com/example/shop/demo/AbusesRepository.java
@@ -1,0 +1,7 @@
+package com.example.shop.demo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AbusesRepository extends JpaRepository<Abuse, Long> {
+
+}